sequential quadrant gating function

The order of 1d-gating is determined so that the gates better capture the distributions of flow data.

gate_quad_sequential(fr, channels, gFunc, min = NULL, max = NULL, ...)

Arguments

fr
flowFrame

channels
character two channels used for gating

gFunc
the name of the 1d-gating function to be used for either dimension

min
a numeric vector that sets the lower bounds for data filtering

max
a numeric vector that sets the upper bounds for data filtering

...
other arguments passed to .find_peak (e.g. 'num_peaks' and 'adjust'). see tailgate

Value

a filters that contains four rectangleGates