generate a partially complete csv template from the existing gating hierarchy

To ease the process of replicating the existing (usually a manual one) gating schemes, this function populate an empty gating template with the 'alias', 'pop', 'parent' and 'dims' columns that exacted from an GatingHierarchy, and leave the other columns (e.g. `gating_method`) blank. So users can make changes to that template instead of writing from scratch.

gh_generate_template(gh)

Arguments

gh
a GatingHierarchy likely parsed from a xml workspace

Value

a gating template in data.frame format that requires further edition after output to csv

Examples

library(flowWorkspace)

#> As part of improvements to flowWorkspace, some behavior of
#> GatingSet objects has changed. For details, please read the section
#> titled "The cytoframe and cytoset classes" in the package vignette:
#> 
#>   vignette("flowWorkspace-Introduction", "flowWorkspace")

dataDir <- system.file("extdata",package="flowWorkspaceData")
gs <- load_gs(list.files(dataDir, pattern = "gs_manual",full = TRUE))
gh_generate_template(gs[[1]])

#>          alias pop                        parent              dims
#> 1   not debris   +                          root       FSC-A,SSC-A
#> 2     singlets   +                   /not debris       FSC-A,FSC-H
#> 3         CD3+   +          /not debris/singlets    <V450-A>,SSC-A
#> 4          DPT   +     /not debris/singlets/CD3+ <B710-A>,<R780-A>
#> 5          DNT   +     /not debris/singlets/CD3+ <B710-A>,<R780-A>
#> 6          CD8   +     /not debris/singlets/CD3+ <B710-A>,<R780-A>
#> 7  CCR7- 45RA-   + /not debris/singlets/CD3+/CD8 <G560-A>,<G780-A>
#> 8  CCR7+ 45RA-   + /not debris/singlets/CD3+/CD8 <G560-A>,<G780-A>
#> 9  CCR7+ 45RA+   + /not debris/singlets/CD3+/CD8 <G560-A>,<G780-A>
#> 10 CCR7- 45RA+   + /not debris/singlets/CD3+/CD8 <G560-A>,<G780-A>
#> 11     38- DR-   + /not debris/singlets/CD3+/CD8 <R660-A>,<V545-A>
#> 12     38+ DR-   + /not debris/singlets/CD3+/CD8 <R660-A>,<V545-A>
#> 13     38+ DR+   + /not debris/singlets/CD3+/CD8 <R660-A>,<V545-A>
#> 14     38- DR+   + /not debris/singlets/CD3+/CD8 <R660-A>,<V545-A>
#> 15         CD4   +     /not debris/singlets/CD3+ <B710-A>,<R780-A>
#> 16 CCR7- 45RA-   + /not debris/singlets/CD3+/CD4 <G560-A>,<G780-A>
#> 17 CCR7+ 45RA-   + /not debris/singlets/CD3+/CD4 <G560-A>,<G780-A>
#> 18 CCR7+ 45RA+   + /not debris/singlets/CD3+/CD4 <G560-A>,<G780-A>
#> 19 CCR7- 45RA+   + /not debris/singlets/CD3+/CD4 <G560-A>,<G780-A>
#> 20     38- DR-   + /not debris/singlets/CD3+/CD4 <R660-A>,<V545-A>
#> 21     38+ DR-   + /not debris/singlets/CD3+/CD4 <R660-A>,<V545-A>
#> 22     38+ DR+   + /not debris/singlets/CD3+/CD4 <R660-A>,<V545-A>
#> 23     38- DR+   + /not debris/singlets/CD3+/CD4 <R660-A>,<V545-A>
#>    gating_method gating_args collapseDataForGating groupBy preprocessing_method
#> 1           <NA>        <NA>                  <NA>    <NA>                 <NA>
#> 2           <NA>        <NA>                  <NA>    <NA>                 <NA>
#> 3           <NA>        <NA>                  <NA>    <NA>                 <NA>
#> 4           <NA>        <NA>                  <NA>    <NA>                 <NA>
#> 5           <NA>        <NA>                  <NA>    <NA>                 <NA>
#> 6           <NA>        <NA>                  <NA>    <NA>                 <NA>
#> 7           <NA>        <NA>                  <NA>    <NA>                 <NA>
#> 8           <NA>        <NA>                  <NA>    <NA>                 <NA>
#> 9           <NA>        <NA>                  <NA>    <NA>                 <NA>
#> 10          <NA>        <NA>                  <NA>    <NA>                 <NA>
#> 11          <NA>        <NA>                  <NA>    <NA>                 <NA>
#> 12          <NA>        <NA>                  <NA>    <NA>                 <NA>
#> 13          <NA>        <NA>                  <NA>    <NA>                 <NA>
#> 14          <NA>        <NA>                  <NA>    <NA>                 <NA>
#> 15          <NA>        <NA>                  <NA>    <NA>                 <NA>
#> 16          <NA>        <NA>                  <NA>    <NA>                 <NA>
#> 17          <NA>        <NA>                  <NA>    <NA>                 <NA>
#> 18          <NA>        <NA>                  <NA>    <NA>                 <NA>
#> 19          <NA>        <NA>                  <NA>    <NA>                 <NA>
#> 20          <NA>        <NA>                  <NA>    <NA>                 <NA>
#> 21          <NA>        <NA>                  <NA>    <NA>                 <NA>
#> 22          <NA>        <NA>                  <NA>    <NA>                 <NA>
#> 23          <NA>        <NA>                  <NA>    <NA>                 <NA>
#>    preprocessing_args
#> 1                <NA>
#> 2                <NA>
#> 3                <NA>
#> 4                <NA>
#> 5                <NA>
#> 6                <NA>
#> 7                <NA>
#> 8                <NA>
#> 9                <NA>
#> 10               <NA>
#> 11               <NA>
#> 12               <NA>
#> 13               <NA>
#> 14               <NA>
#> 15               <NA>
#> 16               <NA>
#> 17               <NA>
#> 18               <NA>
#> 19               <NA>
#> 20               <NA>
#> 21               <NA>
#> 22               <NA>
#> 23               <NA>