apply a gating method to the GatingSet

When interacting with the existing gated data, this function provides an alternative way to interact with the GatingSet by supplying the gating description directly through arguments without the need to write the complete csv gating template.

gs_add_gating_method(gs, alias = "*", pop = "+", parent, dims = NA,
  gating_method, gating_args = NA, collapseDataForGating = NA,
  groupBy = NA, preprocessing_method = NA, preprocessing_args = NA,
  strip_extra_quotes = FALSE, ...)

Arguments

gs
GatingSet or GatingSetList

alias, pop, parent, dims, gating_method, gating_args, collapseDataForGating, groupBy, preprocessing_method, preprocessing_args
see details in gatingTemplate

strip_extra_quotes
logical Extra quotes are added to strings by fread. This causes problems with parsing R strings to expressions in some cases. Default FALSE for usual behaviour. TRUE should be passed if parsing gating_args fails.

...
other arguments

mc.cores passed to multicore package for parallel computing

parallel_type character specifying the parallel type. The valid options are "none", "multicore", "cluster".

cl cluster object passed to parallel package (when parallel_type is "cluster")

Details

Calls to gs_add_gating_method can also be easily reversed with gs_remove_gating_method. Note, however, that it is not possible to differentiate between different GatingSet objects loaded from the same directory with load_gs within a session. Thus, to guarantee a clean history for gs_remove_gating_method, it is necessary to call gs_add_gating_method_init on the loaded GatingSet immediately after re-loading it. See the documentation for gs_add_gating_method_init for more details. This will not be an issue for GatingSet objects created directly using the constructor.

See also

Examples

if (FALSE) {
# add quad gates 
gs_add_gating_method(gs, gating_method = "mindensity", dims = "CCR7,CD45RA", parent = "cd4-cd8+", pop = "CCR7+/-CD45RA+/-")

# polyfunctional gates (boolean combinations of exsiting marginal gates)
gs_add_gating_method(gs, gating_method = "polyFunctions", parent = "cd8", gating_args = "cd8/IFNg:cd8/IL2:cd8/TNFa")

#boolGate method
gs_add_gating_method(gs, alias = "IL2orIFNg", gating_method = "boolGate", parent = "cd4", gating_args = "cd4/IL2|cd4/IFNg")
}