Applies a gatingTemplate to a GatingSet.

It loads the gating methods by topological order and applies them to GatingSet.

gt_gating(x, y, ...)

Arguments

x
a gatingTemplate object

y
a GatingSet object

...

start a character that specifies the population (correspoding to 'alias' column in csv template) where the gating process will start from. It is useful to quickly skip some gates and go directly to the target population in the testing run. Default is "root".

stop.at a character that specifies the population (correspoding to 'alias' column in csv template) where the gating prcoess will stop at. Default is NULL, indicating the end of gating tree.

keep.helperGatesa logical flag indicating whether to keep the intermediate helper gates that are automatically generated by openCyto. Default is TRUE.

mc.cores passed to multicore package for parallel computing

parallel_type character specifying the parallel type. The valid options are "none", "multicore", "cluster".

cl cluster object passed to parallel package (when parallel_type is "cluster")

env_fct
a environment that contains fcTree object named as 'fct'. If NULL (by default), no fcTree will be constructed. It is currently reserved for the internal debugging.

Value

Nothing. As the side effect, gates generated by gating methods are saved in GatingSet.

Examples

if (FALSE) {
 gt <- gatingTemplate(file.path(path, "data/ICStemplate.csv"), "ICS")
 gs <- GatingSet(fs) #fs is a flowSet/ncdfFlowSet
 gt_gating(gt, gs)
 gt_gating(gt, gs, stop.at = "v") #proceed the gating until population 'v'
 gt_gating(gt, gs, start = "v") # start from 'v'
 gt_gating(gt, gs, parallel_type = "multicore", mc.cores = 8) #parallel gating using multicore
 #parallel gating by using cluster
 cl1 <- makeCluster (8, type = "MPI")
 gt_gating(gt, gs, parallel_type = "cluster", cl = cl1)
 stopCluster ( cl1 )
}