Register a gating or preprocessing function with OpenCyto

Function registers a new gating or preprocessing method with openCyto so that it may be used in the csv template.

register_plugins(fun = NA, methodName, dep = NA, ...)

Arguments

fun
function to be registered

methodName
character name of the gating or preprocessing method

dep
character name of the library dependency required for the plugin method to work.

...
other arguments type character specifying the type of registering method. Should be either "gating" or "preprocessing".

Value

logical TRUE if successful and prints a message. FALSE otherwise.

Details

The fun argument should be a wrapper function definition for the gating or preprocessing method. Gating method must have formal arguments:

fr a flowFrame

pp_res a pre-processing result

xChannel character (optional)

yChannel character (required)

filterId character

... ellipses for the additional parameters.

Preprocessing method must have formal arguments:

fs a flowSet that stores the flow data (could be subgrouped data if groupBy column is defined in the csv template

gs a GatingSet

gm a gtMethod object that stores the information from gating method

xChannel character (required)

yChannel character (required)

... ellipses for the additional parameters.

The gating function must return a filter (i.e. polygonGate or other instance) from flowCore. The preprocessing can return anything and it will be passed on to the gating function. So it is up to gating function to use and interpret the results of preprocessing. Not all formal parameters need to be used. Additional arguments are passed via the ... and can be processed in the wrapper